requirements.txt
# Wallpaper Calendar App Dependencies
customtkinter>=5.2.0
Pillow>=10.0.0
tkcalendar>=1.6.1
